pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["juliya-razumova <juliya.razumova@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9"
pandas = "^2.1.4"
numpy = "^1.26.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.8.2"

[tool.poetry.scripts]
main="main:run_all"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
